third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/tables/hidden-attr-expected.txt

This is a testharness.js-based test.
[FAIL] Computed display and visibility of colgroup
  assert_equals: expected "table-column-group" but got "none"
[FAIL] Computed display and visibility of col
  assert_equals: expected "table-column" but got "none"
[FAIL] Computed display and visibility of thead
  assert_equals: expected "table-header-group" but got "none"
[FAIL] Computed display and visibility of tbody
  assert_equals: expected "table-row-group" but got "none"
[FAIL] Computed display and visibility of tfoot
  assert_equals: expected "table-footer-group" but got "none"
[FAIL] Computed display and visibility of tr
  assert_equals: expected "table-row" but got "none"
Harness: the test ran to completion.
